+++ projects/lunaui-angular/src/lib/directives/formatter/public-api.ts
@@ -1,0 +1,2 @@
+export * from './formatter.directive';
+export * from './formatter';