+++ projects/lunaui-angular/src/lib/forms/input/input.ts
@@ -1,0 +1,2 @@
+export type LunaInputSize = 'small' | 'medium' | 'large'
+export type LunaInputVariant = 'filled' | 'outlined' | 'underlined'