+++ projects/lunaui-angular/src/lib/shared/luna-template/public-api.ts
@@ -1,0 +1,1 @@
+export * from './luna-template.directive'