+++ projects/lunaui-angular/src/public-api.ts
@@ -1,5 +1,9 @@
 /*
  * Public API Surface of brickui-angular
 */
-export * from './lib/buttons/public-api'
-export * from './lib/feedback/public-api'
+export * from './lib/buttons/public-api';
+export * from './lib/feedback/public-api';
+export * from './lib/forms/public-api';
+export * from './lib/shared/public-api';
+export * from './lib/directives/public-api';
+export * from './lib/helpers/public-api';